Test fix for macos implementation

+++ mlcolvar/core/loss/utils/smart_derivatives.py
@@ -322,8 +322,8 @@ class SmartDerivatives(torch.nn.Module):
             scatter_min = torch.full((num_groups,), 1e8, device=x.device, dtype=torch.long)
             scatter_max = torch.full((num_groups,), -1e8, device=x.device, dtype=torch.long)
 
-            scatter_min.scatter_reduce_(0, indeces, scatter_indeces, reduce='amin', include_self=True)
-            scatter_max.scatter_reduce_(0, indeces, scatter_indeces, reduce='amax', include_self=True)
+            scatter_min.scatter_reduce_(0, indeces, scatter_indeces.to(torch.int32), reduce='amin', include_self=True)
+            scatter_max.scatter_reduce_(0, indeces, scatter_indeces.to(torch.int32), reduce='amax', include_self=True)
 
             # Compute group spans
             group_spans = (scatter_max - scatter_min + 1)
